Updated partitioning to support a range of USB devices
I encountered issues when attempting to use a 16GB USB stick, with 4K sectors.
Although there is no minimum requirements in the UEFI spec, mkfs.vfat will complain if the partition size is smaller than $sector_size * 65527.
For 512b sector size devices, this is a minimum partition size of ~32MB
For 4K sector size devices, this is a minimum partition size of ~260MB
Some headroom was needed, which is why the recommended size for universal compatibility is 270MB.

Also, updated some partition names and formats.